feat: add PostCast method and organize types
- Add PostCast method to create new Farcaster casts
- Move all type definitions to dedicated types.go file
- Add proper struct definitions for Cast-related types
- Organize imports and dependencies

The PostCast method enables posting casts with support for text content,
embeds, parent references and channel specification.
