+++ main.go
@@ -13,39 +13,6 @@ import (
 	"github.com/ethereum/go-ethereum/crypto"
 )
 
-// LocalAccount represents a local wallet account
-type LocalAccount struct {
-	PrivateKey string
-	PublicKey  string
-	Address    string
-}
-
-// Warpcast represents a client for interacting with the Farcaster API
-type Warpcast struct {
-	config           *ConfigurationParams
-	wallet           *LocalAccount
-	accessToken      *string
-	expiresAt        *int64
-	rotationDuration int64
-	client           *http.Client
-	baseHeaders      map[string]string
-	BaseURL          string
-	HTTPClient       *http.Client
-}
-
-// ConfigurationParams holds the configuration for the Warpcast client
-type ConfigurationParams struct {
-	BasePath    string
-	BaseOptions map[string]interface{}
-	Username    *string
-}
-
-// AuthParams represents the parameters needed for authentication
-type AuthParams struct {
-	Timestamp int64 `json:"timestamp"`
-	// Add other required fields based on the API documentation
-}
-
 // NewWarpcast creates a new Warpcast client instance
 func NewWarpcast(opts ...WarpcastOption) (*Warpcast, error) {
 	w := &Warpcast{
@@ -207,34 +174,6 @@ func (w *Warpcast) createNewAuthToken(duration int64) error {
 	return nil
 }
 
-// AssetResult represents the result of an asset query
-type AssetResult struct {
-	// Add fields as needed
-}
-
-// TokenResult represents the result of a token operation
-type TokenResult struct {
-	Token struct {
-		Secret string `json:"secret"`
-	} `json:"token"`
-}
-
-// StatusContent represents a status response
-type StatusContent struct {
-	Success bool `json:"success"`
-}
-
-// Event represents a single asset event
-type Event struct {
-	// Add relevant fields based on your API response
-}
-
-// IterableEventsResult represents a paginated list of events
-type IterableEventsResult struct {
-	Events []Event
-	Cursor *string
-}
-
 // GetAsset retrieves asset information
 func (w *Warpcast) GetAsset(tokenID int) (*AssetResult, error) {
 	params := map[string]string{
@@ -387,15 +326,6 @@ func (w *Warpcast) generateCustodyAuthHeader(authParams *AuthParams) (string, er
 
 	// Create the custody authorization header
 	return fmt.Sprintf("Bearer %s", signatureHex), nil
-}
-
-// ReactionsPutResult represents the result of liking a cast
-type ReactionsPutResult struct {
-	Like struct {
-		CastHash   string `json:"castHash"`
-		ReactorFid int    `json:"reactorFid"`
-		Timestamp  int64  `json:"timestamp"`
-	} `json:"like"`
 }
 
 // LikeCast likes a given cast
